% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateModel.R
\name{CreateModel}
\alias{CreateModel}
\title{Fits a GLLVM to data}
\usage{
CreateModel(
  abund,
  traits,
  env,
  n_latent = 1,
  FixOmega = NULL,
  FixB = NULL,
  ...
)
}
\arguments{
\item{abund}{Matrix of site by species abundance data}

\item{traits}{Matrix of species by trait data of species traits.}

\item{env}{Matrix of site by environmental variable data.}

\item{n_latent}{The number of latent variables required. Defaults to 1}

\item{FixOmega}{Either NULL (the default, or the index of the value of Omega to be forced to be positive)}

\item{FixB}{Either NULL (the default, or the index of the value of B to be forced to be positive)}

\item{...}{More arguments (none used at the moment)}
}
\value{
A greta model object
}
\description{
Fits a GLLVM to data
}
